Finished memepost app
I completed the model for memepost, and I addeed it to the admin page. I finish the app by making a model have attrubutes of pics, title, and date. I also have Listview to query the latest 25 pictures. I added a tab for mems, but the page is not working correctly. The next commit should allow for ima